Allow enabling bfd protocol conditionally
Why:
The entire rationale is described in projectcalico/calico issue #4607
but the TL;DR is that for allowing quick recovery of BGP kubernetes
service ips in the case of a complete node failure, we need to allow BFD
in calico's bird component.

What:
Add 2 if clause statements, one for enabling the bfd protocol with some
defaults and one for enabling bfd in the bgp_template. The trigger for
enabling this is an environmental variable called ENABLE_BFD