Skip tap/qbr/qvo ports
The tap/qbr/qvo ports are not needed and they are local only ports that are used
inside OpenStack for wiring up VMs.  In hypervisors with many VMs that are also
running Calico (not Calico for OpenStack, just a Kubernetes cluster that uses is
on the same infra), the CPU usage becomes very high for bird because of this.